Backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
transformers==4.35.2
torch==2.1.1
pypdf==3.17.0
google-generativeai==0.3.2
elevenlabs==0.2.26
firebase-admin==6.2.0
python-multipart==0.0.6